Small change to hide API key
