--- client/.env
@@ -1,1 +1,0 @@
-REACT_APP_MAPS_KEY="AIzaSyAeyRbNMuUdpdvj8bd5CTx90-vOrVl2flM"
